requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
boto3==1.38.19
botocore==1.38.19
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.9.0
django-cors-headers==4.7.0
django-currentuser==0.9.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-sequences==3.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
ngrok==1.4.0
oauthlib==3.2.2
openai==1.79.0
paypalrestsdk==1.13.3
phonenumbers==9.0.5
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.0
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stripe==12.2.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
